Refactor main file into multiple files
